core: refactored blockchain.go to blockchain_reader.go #23735
